Crate chapter1

Problems in The Algorithm Design Manual, by S. Skiena, 3rd ed., chapter 1.

The use of traits in this crate is not idiomatic; In a real library, the associated functions would’ve probably been free functions taking in some type implementing a trait providing information on any graph DS.

The goal is to simply group under a single umbrella the methods required to implement a certain algorithm for a specific instance of a specific problem.

Macros

build_error 🔒

ensure_or 🔒

geomatrix 🔒

matrix 🔒

Structs

AdjacencyList 🔒

AdjacencyMatrix 🔒

AdjacencyMatrixError 🔒

Dfs 🔒

GeoAdjacencyMatrix 🔒

Pairs 🔒

PairsError 🔒

Point2d 🔒

Enums

AdjacencyMatrixErrorType 🔒

Edge 🔒

GeoEdge 🔒

PairsErrorType 🔒

Traits

TspClosestPair 🔒

TspMstDfs 🔒

TspNearestNeighbor 🔒